+++ app/smartbrain_3000/web/_app/immutable/entry/start.eRROSM_H.js
@@ -1,0 +1,1 @@
+import{a as e,n as t}from"../chunks/BkQnWF8b.js";export{e as load_css,t as start};